+++ src/hyperion-drm/DRMWrapper.h
@@ -12,7 +12,7 @@ class DRMWrapper : public QObject
 	Q_OBJECT
 public:
 	explicit DRMWrapper(int updateRate_Hz = GrabberWrapper::DEFAULT_RATE_HZ,
-						int display = 0,
+						int deviceIdx = 0,
 						int pixelDecimation = GrabberWrapper::DEFAULT_PIXELDECIMATION,
 						int cropLeft = 0, int cropRight = 0,
 						int cropTop = 0, int cropBottom = 0);