+++ src/hyperion-osx/OsxGrabberCli.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <QCoreApplication>
+
+#include "OsxGrabberOptions.h"
+
+OsxGrabberOptions parseOsxGrabberOptions(const QCoreApplication& app);