+++ src/hyperion-qt/QtGrabberTraits.cpp
@@ -2,20 +2,12 @@
 
 #include <QCoreApplication>
 
-#ifdef ENABLE_MDNS
-#include <mdns/MdnsBrowser.h>
-#else
-#include <ssdp/SSDPDiscover.h>
-#endif
-
-#include <flatbufserver/FlatBufferConnection.h>
+#include <grabberapp/GrabberRunner.h>
 #include <utils/ErrorManager.h>
 #include <utils/Logger.h>
-#include <utils/NetUtils.h>
 
 #include "QtGrabberCli.h"
 #include "QtWrapper.h"
-#include "ScreenshotUtil.h"
 
 void QtGrabberTraits::handleError(QSharedPointer<Logger> log, const QString& error)
 {
@@ -32,6 +24,7 @@ int QtGrabberTraits::run(QCoreApplication& /*app*/,
                          QSharedPointer<Logger> log,
                          ErrorManager& errorManager)
 {
+	// Qt-Grabber specific: the wrapper additionally takes the display index to capture
 	QtWrapper grabber(
 		opts.fps,
 		opts.display,
@@ -41,88 +34,5 @@ int QtGrabberTraits::run(QCoreApplication& /*app*/,
 		opts.cropTop,
 		opts.cropBottom);
 
-	if (!grabber.screenInit())
-	{
-		emit errorManager.errorOccurred(QStringLiteral("Failed to initialise the screen/display for this grabber"));
-		return 1;
-	}
-
-	// set 3D mode if applicable
-	if (opts.video3DSBS)
-	{
-		grabber.setVideoMode(VideoMode::VIDEO_3DSBS);
-	}
-	else if (opts.video3DTAB)
-	{
-		grabber.setVideoMode(VideoMode::VIDEO_3DTAB);
-	}
-
-	if (opts.screenshot)
-	{
-		// Capture a single screenshot and finish
-		const Image<ColorRgb>& screenshot = grabber.getScreenshot();
-		auto const fileName = QStringLiteral("screenshot.png");
-		saveScreenshot(fileName, screenshot);
-		Info(log, "Screenshot saved as: \"%s\"", QSTRING_CSTR(fileName));
-		return 0;
-	}
-
-	QString hostName;
-	int port{FLATBUFFER_DEFAULT_PORT};
-
-	// Split hostname and port (or use default port)
-	QString const givenAddress = opts.address;
-
-	if (!NetUtils::resolveHostPort(givenAddress, hostName, port))
-	{
-		emit errorManager.errorOccurred(QString("Wrong address: unable to parse address (%1)").arg(givenAddress));
-		return 1;
-	}
-
-	Info(log, "Connecting to Hyperion host: %s, port: %u", QSTRING_CSTR(hostName), port);
-
-#ifdef ENABLE_MDNS
-	if (MdnsBrowser::isMdns(hostName))
-	{
-		NetUtils::discoverMdnsServices("flatbuffer");
-	}
-#endif
-
-	if (!NetUtils::convertMdnsToIp(log, hostName, port))
-	{
-		emit errorManager.errorOccurred(QString("IP-address cannot be resolved for the given mDNS service- or hostname: \"%1\"").arg(QSTRING_CSTR(hostName)));
-		return 1;
-	}
-
-	// Create the FlatBuffer-connection
-	FlatBufferConnection const flatbuf(
-		QStringLiteral("Qt-Grabber Standalone"),
-		hostName,
-		opts.priority,
-		opts.skipReply,
-		static_cast<quint16>(port));
-
-	// Connect the screen capturing to flatbuf connection processing
-	QObject::connect(&grabber, &QtWrapper::sig_screenshot,
-	                 &flatbuf,
-	                 static_cast<void (FlatBufferConnection::*)(const Image<ColorRgb>&)>(&FlatBufferConnection::setImage));
-
-	QObject::connect(&flatbuf, &FlatBufferConnection::isReadyToSend, [&log, &grabber]() {
-		Debug(log, "Start grabber");
-		grabber.start();
-	});
-
-	QObject::connect(&flatbuf, &FlatBufferConnection::isDisconnected, [&log, &grabber]() {
-		Debug(log, "Stop grabber");
-		grabber.stop();
-	});
-
-	QObject::connect(&flatbuf, &FlatBufferConnection::errorOccured, [&log, &grabber, &errorManager](const QString& error) {
-		Debug(log, "Stop grabber");
-		grabber.stop();
-		emit errorManager.errorOccurred(error);
-	});
-
-	// Start the application
-	return QCoreApplication::exec();
+	return runFlatbufferScreenGrabber(QString::fromUtf8(Name), grabber, opts, log, errorManager);
 }