+++ src/hyperion-qt/ScreenshotUtil.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <QString>
-
-#include <utils/ColorRgb.h>
-#include <utils/Image.h>
-
-void saveScreenshot(const QString& filename, const Image<ColorRgb>& image);