+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ProviderRestAPI - Add error when failing to load Qt SSL
 - NetUtils: Improve handling when ENABLE_MDNS is false
 - Configure ccache or buildcache only if explicitly requested
+- Standalone grabber apps refactored to a shared Options/Cli/Traits template pattern
 
 ## [2.2.1](https://github.com/hyperion-project/hyperion.ng/releases/tag/2.2.1) - 2026-04-06